requirements/install-py3.txt
attrdict
cerberus>=1.0.1
click
MechanicalSoup
requests
six
websocket-client
